Bug: 385190204
Merged-In: Ia6d2c5c3dd026f2eb66d45027eb2706adf09cef7

+++ native/libimsstack/enabler/mtc/helper/sipinterfaceholder/SessionInterfaceHolder.cpp
@@ -16,15 +16,19 @@
 
 #include "ICoreService.h"
 #include "ISession.h"
+#include "ITimer.h"
+#include "ImsList.h"
 #include "ServiceTrace.h"
+#include "call/IMtcCall.h"
 #include "helper/sipinterfaceholder/IInterfaceHolderListener.h"
 #include "helper/sipinterfaceholder/SessionInterfaceHolder.h"
+#include <unordered_map>
 
 __IMS_TRACE_TAG_COM_MTC__;
 
 PUBLIC
-SessionInterfaceHolder::SessionInterfaceHolder(IN IInterfaceHolderListener& objListener) :
-        m_objListener(objListener)
+SessionInterfaceHolder::SessionInterfaceHolder() :
+        m_objListeners(ImsList<IInterfaceHolderListener*>())
 {
     IMS_TRACE_D("+SessionInterfaceHolder", 0, 0, 0);
 }
@@ -34,45 +38,77 @@ SessionInterfaceHolder::~SessionInterfaceHolder()
 {
     IMS_TRACE_D("~SessionInterfaceHolder", 0, 0, 0);
 
-    ClearISessions();
-
-    for (IMS_SINT32 i = static_cast<IMS_SINT32>(m_objTerminatedGuardTimers.GetSize()) - 1; i >= 0;
-            i--)
+    for (auto& sessionRecord : m_objSessionRecords)
     {
-        StopTimer(m_objTerminatedGuardTimers.GetKeyAt(i));
+        if (sessionRecord.second->piSession != IMS_NULL)
+        {
+            sessionRecord.second->piSession->Destroy();
+        }
+        StopTimer(sessionRecord.second->piTimer);
+        delete sessionRecord.second;
     }
-    m_objTerminatedGuardTimers.Clear();
+
+    m_objSessionRecords.clear();
+}
+
+PUBLIC VIRTUAL void SessionInterfaceHolder::SessionStarted(IN ISession* piSession)
+{
+    IMS_TRACE_D("SessionStarted", 0, 0, 0);
+    piSession->SendAck();
+    piSession->Terminate();
+}
+
+PUBLIC VIRTUAL void SessionInterfaceHolder::SessionStartFailed(IN ISession* piSession)
+{
+    IMS_TRACE_D("SessionStartFailed", 0, 0, 0);
+    ReleaseISession(piSession, IMS_TRUE, IMS_TRUE);
 }
 
 PUBLIC VIRTUAL void SessionInterfaceHolder::SessionTerminated(IN ISession* piSession)
 {
     IMS_TRACE_D("SessionTerminated", 0, 0, 0);
-
-    ReleaseISession(piSession, IMS_TRUE);
+    ReleaseISession(piSession, IMS_TRUE, IMS_TRUE);
 }
 
 PUBLIC VIRTUAL void SessionInterfaceHolder::Timer_TimerExpired(IN ITimer* piTimer)
 {
     IMS_TRACE_D("Timer_TimerExpired", 0, 0, 0);
 
-    IMS_SLONG nIndex = m_objTerminatedGuardTimers.GetIndexOfKey(piTimer);
-    if (nIndex < 0)
+    for (auto pRecord = m_objSessionRecords.begin(); pRecord != m_objSessionRecords.end();
+            ++pRecord)
     {
-        return;
-    }
-
-    ISession* piSession = m_objTerminatedGuardTimers.GetValueAt(nIndex);
-
-    if (piSession)
-    {
-        StopTimer(piTimer);
-        ReleaseISession(piSession, IMS_TRUE);
+        if (pRecord->second->piTimer == piTimer)
+        {
+            ISession* piSession = pRecord->second->piSession;
+            StopTimer(piTimer);
+            ReleaseISession(piSession, IMS_TRUE, IMS_FALSE);
+            break;
+        }
     }
 }
 
 PUBLIC
-ISession* SessionInterfaceHolder::GetISession(
-        IN ICoreService* piCoreService, IN const AString& strFrom, IN const AString& strTo)
+void SessionInterfaceHolder::AddListener(IN IInterfaceHolderListener* piListener)
+{
+    m_objListeners.Append(piListener);
+}
+
+PUBLIC
+void SessionInterfaceHolder::RemoveListener(IN IInterfaceHolderListener* piListener)
+{
+    for (IMS_UINT32 i = 0; i < m_objListeners.GetSize(); i++)
+    {
+        if (m_objListeners.GetAt(i) == piListener)
+        {
+            m_objListeners.RemoveAt(i);
+            return;
+        }
+    }
+}
+
+PUBLIC
+ISession* SessionInterfaceHolder::GetISession(IN CallKey nKey, IN ICoreService* piCoreService,
+        IN const AString& strFrom, IN const AString& strTo)
 {
     IMS_TRACE_D("GetISession", 0, 0, 0);
     if (piCoreService == IMS_NULL)
@@ -81,109 +117,110 @@ ISession* SessionInterfaceHolder::GetISession(
     }
 
     ISession* piSession = piCoreService->CreateSession(strFrom, strTo);
-    m_objISessions.Append(piSession);
+    AddISession(nKey, piSession);
     return piSession;
 }
 
 PUBLIC
-void SessionInterfaceHolder::AddISession(IN ISession* piSession)
+void SessionInterfaceHolder::AddISession(CallKey nKey, IN ISession* piSession)
 {
-    IMS_TRACE_D("AddISession", 0, 0, 0);
-    m_objISessions.Append(piSession);
+    IMS_TRACE_D("AddISession [%d]", nKey, 0, 0);
+    SessionRecord* pSessionRecord = new SessionRecord(piSession);
+    m_objSessionRecords.emplace(nKey, pSessionRecord);
 }
 
 PUBLIC
-void SessionInterfaceHolder::ReleaseISession(
-        IN ISession* piSession, IN IMS_BOOL bTerminated /* = IMS_FALSE*/)
+void SessionInterfaceHolder::ReleaseISession(IN ISession* piSession)
+{
+    ReleaseISession(piSession, IMS_FALSE, IMS_FALSE);
+}
+
+PUBLIC
+void SessionInterfaceHolder::ReleaseISession(IN ISession* piSession, IN IMS_BOOL bEnforceDestroy,
+        IN IMS_BOOL bSessionTerminatedOrStartFailed)
 {
     if (piSession == IMS_NULL)
     {
         return;
     }
 
-    IMS_UINT32 nSize = m_objISessions.GetSize();
-    IMS_TRACE_D("ReleaseISession size=[%d]", nSize, 0, 0);
+    IMS_UINT32 nSize = m_objSessionRecords.size();
+    IMS_TRACE_D("ReleaseISession size=[%d] EnforceDestroy[%s] SessionTerminated(%s)", nSize,
+            _TRACE_B_(bEnforceDestroy), _TRACE_B_(bSessionTerminatedOrStartFailed));
 
     piSession->SetListener(this);
 
-    if (bTerminated || IsReadyToDestroy(piSession))
+    if (bEnforceDestroy || IsReadyToDestroy(piSession, bSessionTerminatedOrStartFailed))
     {
-        for (IMS_UINT32 i = 0; i < nSize; i++)
-        {
-            if (m_objISessions.GetAt(i) == piSession)
-            {
-                ITimer* piTempTimer = GetTimer(piSession);
-                StopTimer(piTempTimer);
-
-                piSession->Destroy();
-                m_objISessions.RemoveAt(i);
-                IMS_TRACE_D("ReleaseISession remove index=[%d]", i, 0, 0);
-
-                if (m_objISessions.GetSize() == 0)
+        auto pRecord = std::find_if(m_objSessionRecords.begin(), m_objSessionRecords.end(),
+                [&piSession](const std::pair<CallKey, SessionRecord*>& record)
                 {
-                    m_objListener.OnSessionInterfaceCleared();
+                    return record.second->piSession == piSession;
+                });
+        if (pRecord != m_objSessionRecords.end())
+        {
+            CallKey nKey = pRecord->first;
+            IMS_TRACE_D("ReleaseISession remove call key=[%d]", nKey, 0, 0);
+
+            ITimer* piTempTimer = pRecord->second->piTimer;
+            StopTimer(piTempTimer);
+
+            piSession->Destroy();
+            delete pRecord->second;
+            m_objSessionRecords.erase(pRecord);
+            if (m_objSessionRecords.count(nKey) == 0)
+            {
+                for (IMS_UINT32 i = 0; i < m_objListeners.GetSize(); ++i)
+                {
+                    m_objListeners.GetAt(i)->OnSessionInterfaceReleased(nKey);
                 }
-                break;
             }
+        }
+        else
+        {
+            IMS_TRACE_D("ReleaseISession not found", 0, 0, 0);
         }
     }
     else
     {
-        StartTimer(piSession, TIME_TRANSACTION_TERMINATED_GUARD);
+        StartTimer(piSession);
     }
 }
 
 PRIVATE
-IMS_BOOL SessionInterfaceHolder::IsReadyToDestroy(IN ISession* piSession)
+IMS_BOOL SessionInterfaceHolder::IsReadyToDestroy(
+        IN ISession* piSession, IN IMS_BOOL bSessionTerminatedOrStartFailed)
 {
     IMS_TRACE_D("IsReadyToDestroy [%d]", piSession->GetState(), 0, 0);
-
-    if (piSession->GetState() == ISession::STATE_TERMINATED)
+    if (bSessionTerminatedOrStartFailed)
     {
-        return IMS_TRUE;
+        if (piSession->GetState() != ISession::STATE_TERMINATING)
+        {
+            return IMS_TRUE;
+        }
     }
-
     return IMS_FALSE;
 }
 
 PRIVATE
-void SessionInterfaceHolder::ClearISessions()
+void SessionInterfaceHolder::StartTimer(IN ISession* piSession)
 {
-    IMS_TRACE_D("ClearISessions [%d]", m_objISessions.GetSize(), 0, 0);
+    IMS_TRACE_D("StartTimer", 0, 0, 0);
 
-    for (IMS_UINT32 i = 0; i < m_objISessions.GetSize(); i++)
+    for (auto& sessionRecord : m_objSessionRecords)
     {
-        ISession* piSession = m_objISessions.GetAt(i);
-        if (piSession != IMS_NULL)
+        if (sessionRecord.second->piSession == piSession)
         {
-            piSession->Destroy();
+            if (sessionRecord.second->piTimer)
+            {
+                return;
+            }
+
+            sessionRecord.second->piTimer = TimerService::GetTimerService()->CreateTimer();
+            sessionRecord.second->piTimer->SetTimer(TIME_TRANSACTION_TERMINATED_GUARD, this);
+            return;
         }
     }
-
-    m_objISessions.Clear();
-}
-
-PRIVATE
-IMS_RESULT SessionInterfaceHolder::StartTimer(IN ISession* piSession, IN IMS_SINT32 nDuration)
-{
-    IMS_TRACE_D("StartTimer duration[%d]", nDuration, 0, 0);
-
-    if (nDuration <= 0)
-    {
-        return IMS_FAILURE;
-    }
-
-    ITimer* piTempTimer = GetTimer(piSession);
-    if (piTempTimer)
-    {
-        return IMS_FAILURE;
-    }
-
-    piTempTimer = TimerService::GetTimerService()->CreateTimer();
-    m_objTerminatedGuardTimers.Add(piTempTimer, piSession);
-    piTempTimer->SetTimer(nDuration, this);
-
-    return IMS_SUCCESS;
 }
 
 PRIVATE
@@ -196,24 +233,27 @@ void SessionInterfaceHolder::StopTimer(IN ITimer* piTimer)
         return;
     }
 
-    m_objTerminatedGuardTimers.Remove(piTimer);
-    piTimer->KillTimer();
-    TimerService::GetTimerService()->DestroyTimer(piTimer);
+    for (auto& sessionRecord : m_objSessionRecords)
+    {
+        if (sessionRecord.second->piTimer == piTimer)
+        {
+            sessionRecord.second->piTimer->KillTimer();
+            TimerService::GetTimerService()->DestroyTimer(piTimer);
+            sessionRecord.second->piTimer = IMS_NULL;
+            break;
+        }
+    }
 }
 
 PRIVATE
 ITimer* SessionInterfaceHolder::GetTimer(IN const ISession* piSession) const
 {
-    IMS_UINT32 nSize = m_objTerminatedGuardTimers.GetSize();
-    IMS_TRACE_D("GetTimer size = [%d]", nSize, 0, 0);
-
-    for (IMS_UINT32 i = 0; i < nSize; i++)
+    for (auto& sessionRecord : m_objSessionRecords)
     {
-        if (piSession == m_objTerminatedGuardTimers.GetValueAt(i))
+        if (sessionRecord.second->piSession == piSession)
         {
-            return m_objTerminatedGuardTimers.GetKeyAt(i);
+            return sessionRecord.second->piTimer;
         }
     }
-
     return IMS_NULL;
 }
